Fixed comment creating. Fix date parse. Fixed post deleting.

+++ lib/ui/entity/uiPostEntity.dart
@@ -13,6 +13,7 @@ class UiPostEntity {
     content = json['content'];
     id = json['id'];
     imageUrl = json['image_url'];
+    // date parse
     final List<String> dateTime = json['create_date'].split(' ');
     final List<String> date = dateTime[0].split('-');
     final List<String> minutes = dateTime[1].split('+')[0].split(':');
@@ -22,7 +23,7 @@ class UiPostEntity {
         int.parse(date[2]),
         int.parse(minutes[0]),
         int.parse(minutes[1]),
-        int.parse(minutes[2]));
+        double.parse(minutes[2]).toInt());
     for (final comment in json['comments'])
       comments.add(Comment()..fromJson(comment));
   }
